requirements.txt
asgiref==3.4.1
Django==3.2.13
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
dnspython==2.2.1
PyJWT==1.7.1
pymongo==4.1.1
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.1.1
